requirements.txt
httpx
python-dotenv
onvif-zeep
zeep
lxml
